Move the ProgressReader.
