+++ cmd/jag/commands/device_network.go
@@ -13,11 +13,9 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"os"
 	"sort"
 	"strings"
 	"time"
-	"unicode/utf8"
 
 	"github.com/libp2p/go-reuseport"
 	"github.com/toitware/ubjson"
@@ -176,59 +174,6 @@ func (d DeviceNetwork) ContainerUninstall(ctx context.Context, sdk *SDK, name st
 		return fmt.Errorf("got non-OK from device: %s", res.Status)
 	}
 	return nil
-}
-
-// A Reader based on a byte array that prints a progress bar.
-type ProgressReader struct {
-	b         []byte
-	index     int
-	spinState int
-}
-
-func NewProgressReader(b []byte) *ProgressReader {
-	return &ProgressReader{b, 0, 0}
-}
-
-func (p *ProgressReader) Read(buffer []byte) (n int, err error) {
-	if p.index == len(p.b) {
-		return 0, io.EOF
-	}
-	copied := copy(buffer, p.b[p.index:])
-	p.index += copied
-	percent := (p.index * 100) / len(p.b)
-	fmt.Print("\r")
-	// The strings must contain characters with the same UTF-8 length so that
-	// they can be chopped up.  The emoji generally are 4-byte characters.
-	// Braille are 3-byte characters, and or course ASCII is 1-byte characters.
-	spin := "⠁⠂⠄⡀⢀⠠⠐⠈"
-	done := "🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱🐱"
-	todo := "--------------------------------------------------"
-	if os.PathSeparator == '\\' { // Windows.
-		spin = "/-\\|"
-		done = "################### Jaguar #######################"
-	}
-
-	parts := utf8.RuneCountInString(done)
-	todoParts := utf8.RuneCountInString(todo)
-	if todoParts < parts {
-		parts = todoParts
-	}
-	spinStates := utf8.RuneCountInString(spin)
-	doneBytesPerPart := len(done) / parts
-	todoBytesPerPart := len(todo) / parts
-	spinBytesPerPart := len(spin) / spinStates
-
-	pos := percent / (100 / parts)
-	p.spinState += spinBytesPerPart
-	if p.spinState == len(spin) {
-		p.spinState = 0
-	}
-	spinChar := spin[p.spinState : p.spinState+spinBytesPerPart]
-	fmt.Printf("   %3d%%  %4dk  %s  [", percent, p.index>>10, spinChar)
-	fmt.Print(done[len(done)-pos*doneBytesPerPart:])
-	fmt.Print(todo[:len(todo)-pos*todoBytesPerPart])
-	fmt.Print("] ")
-	return copied, nil
 }
 
 func (d DeviceNetwork) UpdateFirmware(ctx context.Context, sdk *SDK, b []byte) error {
